update examples, and rename api.rb -> flickr.rb & oauth.rb -> oauth_client.rb
